git subrepo commit (merge) c2rust @ a87d45831 (#425)
subrepo:
  subdir:   "c2rust"
  merged:   "1cf803036"
upstream:
  origin:   "git@github.com:Aarno-Labs/c2rust.git"
  branch:   "tenjin"
  commit:   "9a8d10cfd"
git-subrepo:
  version:  "0.4.9"
  origin:   "https://github.com/ingydotnet/git-subrepo"
  commit:   "4f60dd7"

+++ tests/snapshotted/ffi_multifile_guidance/_xj_snapshot/00_out/src/strings.rs
@@ -16,7 +16,7 @@ pub unsafe extern "C" fn strings_demo() -> ::core::ffi::c_int {
         9 as ::core::ffi::c_uchar,
     ];
     zero_first(&mut buf, 4 as ::core::ffi::c_int);
-    return buf[0 as usize] as ::core::ffi::c_int;
+    return buf[0usize] as ::core::ffi::c_int;
 }
 pub mod xj_ffi {
     #[allow(unused_imports)]
