+++ tests/snapshotted/pointer_param_reseat/_xj_snapshot/00_out/pointer_param_reseat/src/pointer_param_reseat.rs
@@ -26,7 +26,7 @@ pub unsafe extern "C" fn write_not_null(mut dest: *mut ::core::ffi::c_int) -> ::
     let c2rust_fresh2 = dest;
     dest = dest.offset(1);
     *c2rust_fresh2 = 3 as ::core::ffi::c_int;
-    return buf[0 as usize] + buf[1 as usize] + buf[2 as usize];
+    return buf[0usize] + buf[1usize] + buf[2usize];
 }
 unsafe fn main_0() -> ::core::ffi::c_int {
     return write_not_null(::core::ptr::null_mut::<::core::ffi::c_int>());
